# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kkiit Marketplace API
  version: 0.2.0
  description: Kkiit 재능·AI 서비스 마켓플레이스 REST API
servers:
  - url: /api/v1
security:
  - cookieSession: []
  - bearerApiKey: []
paths:
...
  /auth/oauth/{slug}/start:
    get:
      security: []
      summary: 소셜·OIDC 로그인 시작
      description: 제공자의 인가 화면으로 리다이렉트한다. `link=1`이면 로그인 대신 현재 계정에 연결하는 흐름으로 진행한다.
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: link, in: query, schema: { type: string, enum: ['1'] } }
      responses:
        '302': { description: redirect to the provider }
        '404': { $ref: '#/components/responses/Error' }
  /auth/oauth/{slug}/callback:
    get:
      security: []
      summary: 소셜·OIDC 로그인 콜백
      description: 제공자가 돌려보낸 code와 state를 검증한다. state는 1회용이며 만료된다.
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        '302': { description: redirect back into the application }
        '400': { $ref: '#/components/responses/Error' }
...
  /auth/login:
    post:
      security: []
      summary: 로컬 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200': { description: login success }
        '401': { $ref: '#/components/responses/Error' }
...
  /me/reports:
    get:
      summary: 내 신고 내역
      parameters: [{ $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: reports with their state and resolution }
  /me/settlements:
    get:
      summary: 내 정산 내역과 수익 요약
      description: 판매자 본인의 정산만 반환하며 보류 사유와 지급 예정일을 포함한다.
      parameters:
        - { name: state, in: query, schema: { type: string, enum: [scheduled, confirmed, hold, completed, cancelled] } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: settlements with next_cursor and an earnings summary }
        '403': { $ref: '#/components/responses/Error' }
  /me/favorites:
    get:
      summary: 찜한 서비스 목록
      parameters: [{ $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: saved talents with seller trust information }
  /me/files:
    post:
      summary: 공개 미디어 업로드
      description: 포트폴리오처럼 방문자에게 보여줄 파일을 공개 표시로 저장한다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: stored file with a download url }
        '413': { $ref: '#/components/responses/Error' }
  /me/portfolios:
    get:
      summary: 내 포트폴리오
      responses:
        '200': { description: portfolios }
    post:
      summary: 포트폴리오 등록
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioInput' }
      responses:
        '201': { description: created }
        '400': { $ref: '#/components/responses/Error' }
  /me/portfolios/{id}:
    put:
      summary: 포트폴리오 수정
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioInput' }
      responses:
        '200': { description: updated }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      summary: 포트폴리오 삭제
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: deleted }
  /me/notifications:
    get:
      summary: 내 알림함
      parameters:
        - { name: unread, in: query, schema: { type: boolean }, description: 읽지 않은 알림만 조회 }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: notifications with unread count }
  /me/notifications/read:
    post:
      summary: 알림 읽음 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids: { type: array, items: { type: string, format: uuid } }
                all: { type: boolean, description: 읽지 않은 알림 전체를 읽음 처리 }
      responses:
        '200': { description: remaining unread count }
        '400': { $ref: '#/components/responses/Error' }
...
  /me/notification-preferences:
    get:
      summary: 이벤트별 알림 수신 설정
      responses:
        '200': { description: preferences for every catalog event }
    put:
      summary: 이벤트별 알림 수신 설정 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [event_type, enabled]
                    properties:
                      event_type: { type: string }
                      channels: { type: array, items: { type: string } }
                      enabled: { type: boolean }
      responses:
        '200': { description: saved }
        '400': { $ref: '#/components/responses/Error' }
  /me/webhooks:
    get:
      summary: 내 이벤트 웹훅 목록
      responses:
        '200': { description: webhooks and the subscribable event catalog }
    post:
      summary: 이벤트 웹훅 등록
      description: 서명 키 원문은 이 응답에서 한 번만 반환된다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookInput' }
      responses:
        '201': { description: created webhook with one-time signing secret }
        '400': { $ref: '#/components/responses/Error' }
  /me/webhooks/{id}:
    put:
      summary: 이벤트 웹훅 변경
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookInput' }
      responses:
        '200': { description: updated, includes a new secret when rotated }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      summary: 이벤트 웹훅 삭제
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: deleted }
  /me/webhooks/{id}/test:
    post:
      summary: 웹훅 테스트 이벤트 전송
      description: 실제 이벤트와 같은 서명·재시도 경로로 `WebhookTest` 이벤트를 큐에 넣는다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '202': { description: queued }
        '409': { $ref: '#/components/responses/Error' }
...
  /me/api-keys/{id}/rotate:
    post:
      summary: API Key 원자적 회전
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: rotated key with one-time secret }
  /me/password:
    post:
      summary: 비밀번호 변경
      description: 현재 비밀번호를 확인하고 교체한 뒤 현재 기기를 제외한 모든 세션을 종료한다. 비밀번호가 없는 소셜 계정은 current_password 없이 설정할 수 있다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 12 }
      responses:
        '200': { description: changed with the number of revoked sessions }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /me/identities/{id}:
    delete:
      summary: 제공자 연결 해제
      description: 계정의 마지막 로그인 수단은 해제할 수 없다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: unlinked }
        '409': { $ref: '#/components/responses/Error' }
  /me/api-keys/{id}:
    delete:
      summary: API 키 폐기
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: key revoked }
        '404': { $ref: '#/components/responses/Error' }
...
  /talents:
    get:
      security: []
      summary: 공개 재능 상품 FTS 검색
      description: 기본 정렬은 search.policy 가중치로 텍스트 적합도, 상품 품질 점수, 최신성을 섞는다. 각 항목은 판매자 등급과 평점을 포함하며, 응답은 `has_more`와 `next_offset`을 함께 돌려준다.
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string }, description: 카테고리 slug. 상위 카테고리를 지정하면 하위 카테고리 상품도 함께 반환한다. }
        - { name: price_min, in: query, schema: { type: integer, minimum: 1 } }
        - { name: price_max, in: query, schema: { type: integer, minimum: 1 } }
        - { name: max_delivery_days, in: query, schema: { type: integer, minimum: 1 } }
        - { name: service_type, in: query, schema: { type: string, enum: [HUMAN, AI, HYBRID] } }
        - { name: sort, in: query, schema: { type: string, enum: [recommended, price_asc, price_desc, delivery, rating, newest] }, description: 지정하지 않으면 추천순. }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 1000 }, description: 순위가 요청마다 계산되므로 커서 대신 offset으로 넘긴다. 1000을 넘는 값은 1000으로 잘린다. }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        '200': { description: talents ordered by the blended rank score }
    post:
      summary: 재능 상품과 패키지, Smart Order Form 등록
      responses:
        '201': { description: draft talent }
  /talents/{id}:
    get:
      security: []
      summary: 재능 상세
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: talent detail }
    put:
      summary: 본인 재능 상품 변경
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: updated }
  /talents/{id}/publish:
    post:
      summary: 재능 공개 또는 승인 요청
      description: 조건과 일치하는 활성 talent_publish 정책이 없으면 즉시 공개한다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: published or review_pending }
...
  /me/reviews:
    get:
      summary: 내가 받은 후기
      description: 내 상품에 달린 후기를 최신순 커서 페이징으로 반환한다. 각 항목에 상품과 주문 번호가 함께 오며, 응답의 `unanswered`는 아직 답글을 남기지 않은 건수다. `unanswered=1`이면 답글 없는 것만 반환한다.
      parameters:
        - { name: unanswered, in: query, schema: { type: string, enum: ['1'] } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: reviews with next_cursor and the unanswered count }
  /reviews/{id}/reply:
    post:
      summary: 리뷰에 대한 판매자 답글
      description: 해당 주문의 판매자만 남길 수 있다. 리뷰 본문과 점수는 바뀌지 않으며 답글은 리뷰 아래에 함께 공개된다. 다시 호출하면 답글을 고친다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: reply saved }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      summary: 판매자 답글 삭제
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: reply removed }
        '403': { $ref: '#/components/responses/Error' }
  /talents/{id}/view:
    post:
      security: []
      summary: 상품 조회 기록
      description: 뷰어당 상품당 하루 한 번만 집계하며 판매자 본인의 조회와 비공개 상품은 세지 않는다. 비로그인 뷰어는 주소를 저장하지 않고 해시로만 구분한다. 실패해도 페이지에는 영향을 주지 않으므로 항상 204를 돌려준다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: recorded or intentionally ignored }
        '404': { $ref: '#/components/responses/Error' }
  /talents/{id}/status:
    post:
      summary: 판매자가 직접 하는 판매 중지·재개
      description: 본인 상품만 대상으로 한다. `published`↔`paused`, `draft`→`archived`, `archived`→`published`/`paused`/`draft`만 허용한다. 중지는 새 주문만 막고 진행 중인 주문은 그대로 둔다. 재개는 이미 승인된 내용이므로 재검토를 거치지 않으며, 내용을 고치면 수정 경로가 다시 검토로 보낸다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: status changed }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /orders:
    get:
      summary: 접근 가능한 구매·판매 주문
      description: 본인이 구매자 또는 판매자인 주문을 최신순 커서 페이징으로 반환한다. `orders.manage` 권한이 있으면 전체 주문을 반환한다. 알 수 없는 `state` 값은 필터 없이 무시한다.
      parameters:
        - { name: role, in: query, schema: { type: string, enum: [buyer, seller] } }
        - { name: state, in: query, schema: { type: string }, description: 주문 상태 또는 `OPEN`(종료되지 않은 모든 주문). }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: orders with next_cursor }
    post:
      summary: 공개 상품 주문
      description: 판매자가 정의한 주문 양식을 서버에서 검증한다. 필수 항목이 비어 있으면 400을 돌려주며, 답변은 요구사항 id 또는 label 어느 쪽으로 보내도 되고 저장은 항상 label 기준이다. 양식에 없는 항목은 저장하지 않는다.
      responses:
        '201': { description: created order }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /orders/{id}:
    get:
      summary: 주문 Workspace 데이터와 Timeline
      description: 구매자·판매자·운영자, 그리고 조직 명의 주문이라면 그 조직의 소유자·관리자가 열람합니다. 납기를 넘긴 미납품 주문은 `overdue`와 `overdue_days`를 함께 반환합니다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: order workspace }
        '404': { $ref: '#/components/responses/Error' }
  /orders/{id}/transition:
    post:
      summary: 허용표 기반 주문 상태 변경
      description: 취소 완료는 원칙적으로 `orders.manage` 권한자의 몫이지만, 납기를 `marketplace.policy.overdue_cancel_days`(기본 7일)만큼 넘긴 미납품 주문은 구매자가 직접 마무리하고 에스크로를 돌려받는다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: transitioned }
        '409': { $ref: '#/components/responses/Error' }
  /orders/{id}/pay:
    post:
      summary: 결제 Adapter 실행과 에스크로 원장 기록
      description: Idempotency-Key 헤더가 필수이며 현재 기본 Adapter는 관리자 정책에 따른 manual 방식이다.
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      responses:
        '200': { description: payment recorded idempotently }
...
  /orders/{id}/revision:
    post:
      summary: 구매자 수정 요청
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: revision }
  /orders/{id}/accept:
    post:
      summary: 구매확정 및 정산 원장 생성
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: accepted and settlement scheduled }
  /orders/{id}/messages:
    get:
      summary: 주문 메시지와 읽음 상태 조회
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: messages }
    post:
      summary: 주문 메시지 전송
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: message created }
  /orders/{id}/messages/ws:
    get:
      summary: 주문 실시간 메시지 WebSocket
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '101': { description: WebSocket upgraded }
  /orders/{id}/files:
    post:
      summary: 주문 파일 업로드와 SHA-256 무결성 기록
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: file metadata and download URL }
  /orders/{id}/review:
    post:
      summary: 완료 주문 다면 리뷰 등록
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: review created }
...
  /talents/{id}/inquiries:
    post:
      summary: 주문 전 상품 문의
      description: 공개된 상품에만 보낼 수 있고 본인 상품에는 보낼 수 없다. 같은 구매자가 같은 상품에 다시 문의하면 새 스레드가 아니라 기존 대화가 이어진다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: inquiry thread }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /me/inquiries:
    get:
      summary: 내 상품 문의 목록
      description: 내가 보낸 문의와 내 상품에 들어온 문의를 최근 대화순으로 함께 반환한다. 각 항목의 `role`이 내 입장, `unread`가 아직 읽지 않은 상대 메시지 수다.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: inquiry threads with next_cursor }
  /inquiries/{id}/messages:
    get:
      summary: 문의 대화 내용
      description: 당사자 두 사람만 볼 수 있다. 열람하면 상대가 보낸 메시지가 읽음으로 표시된다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: messages in the thread }
        '403': { $ref: '#/components/responses/Error' }
    post:
      summary: 문의 답장
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '201': { description: message sent }
        '403': { $ref: '#/components/responses/Error' }
...
  /sellers/{id}/talents:
    get:
      security: []
      summary: 판매자의 공개 상품 목록
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { description: published talents }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: kkiit_session }
    bearerApiKey: { type: http, scheme: bearer }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1 }, description: 반환 최대 건수 }
    Cursor: { name: cursor, in: query, schema: { type: string }, description: 이전 응답의 next_cursor. 만료되거나 잘못된 값은 최신 페이지로 되돌아간다. }
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
  schemas:
...
    PortfolioInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, maxLength: 4000 }
        media:
          type: array
          maxItems: 12
          items:
            type: object
            properties:
              url: { type: string }
              name: { type: string }
              mime_type: { type: string }
        tags: { type: array, maxItems: 20, items: { type: string } }
...